move geometry to common
