+++ common/geometry.cpp
